bcu-downloader/requirements.txt
psycopg2==2.8.4
python-dateutil==2.8.1
selenium==3.141.0
six==1.14.0
urllib3==1.25.8
